refactor(iocuak): move bind utils to container folder
