+++ packages/iocuak/src/container/utils/bindToValue.ts
@@ -1,7 +1,7 @@
 import { ServiceId } from '../../common/models/domain/ServiceId';
-import { ContainerBindingService } from '../../container/services/domain/ContainerBindingService';
 import { BindingType } from '../../metadata/models/domain/BindingType';
 import { ValueBinding } from '../../metadata/models/domain/ValueBinding';
+import { ContainerBindingService } from '../services/domain/ContainerBindingService';
 
 export function bindToValue<TInstance>(
   serviceId: ServiceId,
